+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file. This change
 
 ## [unreleased]
 
+## [0.2.0] - 2017-03-22
+### Added
+- Automatic support for both pymongo and pybson bson codecs
+  (https://pypi.python.org/pypi/pymongo & https://pypi.python.org/pypi/bson/0.4.6)
+- Supports providing custom json codec and custom bson codec implementations via options.
+
 ## [0.1.2] - 2016-04-13
 ### Fixes
 - Fix threading unstability issue related to dispatcher thread starting too early.
@@ -20,7 +26,8 @@ All notable changes to this project will be documented in this file. This change
 ### Initial Version:
 - JSON-RPC 2.0 and BSON-RPC library on sockets/gevent.sockets.
 
-[unreleased]: https://github.com/seprich/py-bson-rpc/compare/0.1.2...HEAD
+[unreleased]: https://github.com/seprich/py-bson-rpc/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/seprich/py-bson-rpc/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/seprich/py-bson-rpc/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/seprich/py-bson-rpc/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/seprich/py-bson-rpc/tree/0.1.0